chore: move trig/const/rand tests to src dir
